fix issue when threshold is using default value
Even though we set settings.awayThreshold in initialize() when its
null, it doesn't seem to affect the global awayThreshold value. Since
we are using that global value in reasonStr(), it is still null if
user did not specify a value. So we see the string "null" in the
notification message.

This fix updates reasonStr() to handle this case. We still want to use
the global awayThreshold/arrivalThreshold value (instead of converting
delaySec by / 60) because when the value is a fraction, conversion will
give us a non exact floating point value, so it's better to use the
value the user specified directly.